core: compile statically
Compiling core dynamically requires relocations which are a real pain. If we choose to ignore relocations, we won't be able to use strings either.
A better solution would be to compile core statically. This is fine in our case. Since core's physical address is constant (by design), its
virtual address is constant from its context as well (it is only executed when waking from suspend).