implement fixed-size blocks allocators
This drastically increases the allocator's space efficiency, since it
can now free unused memory.

This change is crucial for spamming read/write requests, since every
request allocates a `virtio_blk_req`.
On the previous simple allocator, free was a noop, and the request
allocation was never freed. Therefore, there was a limit to the number
of requests possible.