UEFI: add EFI application to reserve memory
Core is loaded into a fixed physical address which must not be touched
by the OS.

The previous implementation passed a `memmap` boot param to the kernel,
which altered the E820 memory mapping passed to the kernel, in order to
reserve this memory.

In order to be more generic (and to allow future switch-os
implementations for more kernels) we do this using a UEFI application.