virtio: consume responses after each request
Previously I've added an optimization that makes all the read/write
requests in parallel, and then consumes all the responses.

This is problematic since it allocates too many queue descriptors at
once, and it ends up filling the descriptors queue.

A more efficient solution would be to create some requests in parallel
while taking the descriptors queue size in consideration.