feature: pass original waking vector to core
The sleep hooks (_PTS and SetVariable) override the waking vector to
point to core, but they also need to store the original waking vector so
core can return back to the kernel when it's done.

Ideally we'd want to directly store the waking vector inside core's
header, but it doesn't seem to be mapped for the AML VM, so the _PTS
hook can't reach it. It might also not be available to the UEFI Runtime
Services environment, but I haven't checked that.

Instead, we can use the second reserved field in the FACS table, which
is 24-bytes long. We use the beginning of this field to store the
original waking vector, and then retrieve it in core.